cairo: rename cairo-xxx.x to xxx.x
