+++ binding.gyp
@@ -18,17 +18,17 @@
                 "src/value.cc",
                 "src/modules/system.cc",
                 "src/modules/cairo/cairo.cc",
-                "src/modules/cairo/cairo-context.cc",
-                "src/modules/cairo/cairo-matrix.cc",
-                "src/modules/cairo/cairo-path.cc",
-                "src/modules/cairo/cairo-pattern.cc",
-                "src/modules/cairo/cairo-font-options.cc",
-                "src/modules/cairo/cairo-font-extents.cc",
-                "src/modules/cairo/cairo-text-extents.cc",
-                "src/modules/cairo/cairo-region.cc",
-                "src/modules/cairo/cairo-rectangle.cc",
-                "src/modules/cairo/cairo-rectangle-int.cc",
-                "src/modules/cairo/cairo-surface.cc",
+                "src/modules/cairo/context.cc",
+                "src/modules/cairo/matrix.cc",
+                "src/modules/cairo/path.cc",
+                "src/modules/cairo/pattern.cc",
+                "src/modules/cairo/font-options.cc",
+                "src/modules/cairo/font-extents.cc",
+                "src/modules/cairo/text-extents.cc",
+                "src/modules/cairo/region.cc",
+                "src/modules/cairo/rectangle.cc",
+                "src/modules/cairo/rectangle-int.cc",
+                "src/modules/cairo/surface.cc",
             ],
             "include_dirs" : [
                 "<!(node -e \"require('nan')\")"
