+++ src/modules/cairo/cairo.cc
@@ -8,17 +8,17 @@
 #include "../../gi.h"
 #include "../../gobject.h"
 #include "../../value.h"
-#include "cairo-context.h"
-#include "cairo-font-options.h"
-#include "cairo-font-extents.h"
-#include "cairo-text-extents.h"
-#include "cairo-matrix.h"
-#include "cairo-path.h"
-#include "cairo-pattern.h"
-#include "cairo-rectangle.h"
-#include "cairo-rectangle-int.h"
-#include "cairo-region.h"
-#include "cairo-surface.h"
+#include "context.h"
+#include "font-options.h"
+#include "font-extents.h"
+#include "text-extents.h"
+#include "matrix.h"
+#include "path.h"
+#include "pattern.h"
+#include "rectangle.h"
+#include "rectangle-int.h"
+#include "region.h"
+#include "surface.h"
 
 using v8::Function;
 using v8::FunctionTemplate;
