+++ src/modules/cairo/pattern.h
